requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.5.2
APScheduler==3.10.4
async-timeout==5.0.1
attrs==24.2.0
backports.zoneinfo==0.2.1
certifi==2024.8.30
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-telegram-bot==20.3
pytz==2024.2
six==1.15.0
sniffio==1.3.1
typing_extensions==4.12.2
tzlocal==5.2
watchdog==4.0.2
yarl==1.15.2
